requirements.txt
matplotlib==3.10.8
pandas==2.1.4
psycopg2-binary==2.9.11
requests==2.31.0
seaborn==0.13.2
